+++ packages/gallery/src/components/gallery/proGallery/galleryContainerNew.js
@@ -1,22 +1,22 @@
 import React from 'react';
 import GalleryView from './galleryView';
 import SlideshowView from './slideshowView';
-import { addLayoutStyles } from '../helpers/layoutHelper';
-import { ItemsHelper } from '../helpers/itemsHelper';
-import dimensionsHelper from '../helpers/dimensionsHelper';
-import { scrollToItemImp } from '../helpers/scrollHelper';
-import window from '../../common/window/windowWrapper';
+import { addLayoutStyles } from '../../helpers/layoutHelper';
+import { ItemsHelper } from '../../helpers/itemsHelper';
+import dimensionsHelper from '../../helpers/dimensionsHelper';
+import { scrollToItemImp } from '../../helpers/scrollHelper';
+import window from '../../../common/window/windowWrapper';
 import ScrollIndicator from './galleryScrollIndicator';
 import { Layouter } from 'pro-layouts';
-import { cssScrollHelper } from '../helpers/cssScrollHelper.js';
-import { createCssLayouts } from '../helpers/cssLayoutsHelper.js';
-import utils from '../../common/utils';
-import { isEditMode, isSEOMode, isPreviewMode, isSiteMode } from '../../common/window/viewModeWrapper';
-import EVENTS from '../../common/constants/events';
-import CLICK_ACTIONS from '../../common/constants/itemClick';
-import VideoScrollHelper from '../helpers/videoScrollHelper.js';
-import { URL_TYPES, URL_SIZES } from '../../common/constants/urlTypes';
-import checkNewGalleryProps from './helpers/isNew';
+import { cssScrollHelper } from '../../helpers/cssScrollHelper.js';
+import { createCssLayouts } from '../../helpers/cssLayoutsHelper.js';
+import utils from '../../../common/utils';
+import { isEditMode, isSEOMode, isPreviewMode, isSiteMode } from '../../../common/window/viewModeWrapper';
+import EVENTS from '../../../common/constants/events';
+import CLICK_ACTIONS from '../../../common/constants/itemClick';
+import VideoScrollHelper from '../../helpers/videoScrollHelper.js';
+import { URL_TYPES, URL_SIZES } from '../../../common/constants/urlTypes';
+import checkNewGalleryProps from '../../helpers/isNew';
 
 export class GalleryContainer extends React.Component {
   constructor(props) {
@@ -617,14 +617,16 @@ export class GalleryContainer extends React.Component {
           durationInMS,
         };
         return scrollToItemImp(scrollParams);
-      } catch {
+      } catch(e) {
         //added console.error to debug sentry error 'Cannot read property 'isRTL' of undefined in pro-gallery-statics'
-        console.error('pro-gallery, scrollToItem, cannot get scrollParams, ',
-          'isEditMode =', isEditMode,
-          ' isPreviewMode =', isPreviewMode,
-          ' isSiteMode =', isSiteMode,
-          ' this.state.styles =', this.state.styles
-          );
+        console.error('error:', e, ' pro-gallery, scrollToItem, cannot get scrollParams, ',
+          'isEditMode =', isEditMode(),
+          ' isPreviewMode =', isPreviewMode(),
+          ' isSiteMode =', isSiteMode(),
+          ' this.state.styles =', this.state.styles,
+          ' this.state.container =', this.state.container,
+          ' this.galleryStructure =', this.galleryStructure
+        );
       }
 
     }