+++ packages/gallery/src/index.js
@@ -1,10 +1,10 @@
 import './components/styles/gallery.scss';
 
-import ProGallery from './components/gallery/proGallery';
+import ProGallery from './components/gallery/index';
 import LoveButton from './components/item/loveButton/loveButton';
 import GalleryItem from './components/item/galleryItem';
 import gallerySettings from './settings';
-import ExpandableProGallery from './components/gallery/withFullscreen';
+import ExpandableProGallery from './components/gallery/presets/expandableGallery';
 import GALLERY_CONSTS from './common/constants';
 
 export {