+++ packages/gallery/tests/drivers/matchers.js
@@ -1,0 +1,8 @@
+import { configureToMatchImageSnapshot } from 'jest-image-snapshot';
+
+export const toMatchImageSnapshot = configureToMatchImageSnapshot({
+  noColors: true,
+  failureThreshold: 1,
+  failureThresholdType: 'percent',
+  blur: 1
+});