+++ packages/gallery/tests/e2e/styleParams/imageMargin.e2e.spec.js
@@ -1,5 +1,5 @@
 import GalleryDriver from '../../drivers/pptrDriver';
-import { toMatchImageSnapshot} from 'jest-image-snapshot';
+import {toMatchImageSnapshot} from '../../drivers/matchers';
 
 jest.setTimeout(30000)
 
@@ -22,6 +22,7 @@ describe('imageMargin - e2e', () => {
       imageMargin: 20
     });
     await driver.waitFor.hookToBeVisible('item-container');
+    await driver.waitFor.timer(2000);
     const page = await driver.grab.elemScreenshot('#pro-gallery-container');
     expect(page).toMatchImageSnapshot();
   });
@@ -31,6 +32,7 @@ describe('imageMargin - e2e', () => {
       imageMargin: 0
     });
     await driver.waitFor.hookToBeVisible('item-container');
+    await driver.waitFor.timer(2000);
     const page = await driver.grab.elemScreenshot('#pro-gallery-container');
     expect(page).toMatchImageSnapshot();
   });