+++ packages/playground/src/settings/dataTypes.js
@@ -4,6 +4,9 @@ const formatTitle = title => title.replace(/_/g, ' ').split(' ').map(word => wor
 const createOptions = constName => Object.entries(GALLERY_CONSTS[constName]).map(([title, value]) => ({ value, title: formatTitle(title) }));
 
 export default {
+  allowLeanGallery: {
+    type: INPUT_TYPES.BOOLEAN,
+  },
   isStore: {
     type: INPUT_TYPES.BOOLEAN,
     alert: 'MOCK PARAM',
@@ -629,6 +632,9 @@ export default {
     max: 255,
     units: 'lv',
     alert: 'inside "sharpParams" object as "usm_t" property + percentage - should be divided by 255',
+  },
+  hidePlay: {
+    type: INPUT_TYPES.BOOLEAN,
   },
   videoPlay: {
     type: INPUT_TYPES.OPTIONS,